+++ modules/library.py
@@ -3,6 +3,7 @@ import time
 from abc import ABC, abstractmethod
 
 from PIL import Image
+from requests.exceptions import RequestException
 
 from modules import timings, util
 from modules.meta import MetadataFile, OverlayFile
@@ -311,7 +312,7 @@ class Library(ABC):
                     logger.info(f"Metadata: {poster.attribute} updated {poster.message}")
                 elif self.show_asset_not_needed:
                     logger.info(f"Metadata: {poster.prefix}poster update not needed")
-            except Failed:
+            except (Failed, RequestException):
                 logger.stacktrace()
                 logger.error(f"Metadata: {poster.attribute} failed to update {poster.message}")
 
@@ -326,7 +327,7 @@ class Library(ABC):
                     logger.info(f"Metadata: {background.attribute} updated {background.message}")
                 elif self.show_asset_not_needed:
                     logger.info(f"Metadata: {background.prefix}background update not needed")
-            except Failed:
+            except (Failed, RequestException):
                 logger.stacktrace()
                 logger.error(f"Metadata: {background.attribute} failed to update {background.message}")
 
@@ -341,7 +342,7 @@ class Library(ABC):
                     logger.info(f"Metadata: {logo.attribute} updated {logo.message}")
                 elif self.show_asset_not_needed:
                     logger.info(f"Metadata: {logo.prefix}logo update not needed")
-            except Failed:
+            except (Failed, RequestException):
                 logger.stacktrace()
                 logger.error(f"Metadata: {logo.attribute} failed to update {logo.message}")
 
@@ -356,7 +357,7 @@ class Library(ABC):
                     logger.info(f"Metadata: {square_art.attribute} updated {square_art.message}")
                 elif self.show_asset_not_needed:
                     logger.info(f"Metadata: {square_art.prefix}square art update not needed")
-            except Failed:
+            except (Failed, RequestException):
                 logger.stacktrace()
                 logger.error(f"Metadata: {square_art.attribute} failed to update {square_art.message}")